Edit folder project

+++ src/main/java/compedia/vn/tickmi/mail/repository/EventRequestRepository.java
@@ -4,6 +4,7 @@ import compedia.vn.tickmi.mail.entity.EventRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface EventRequestRepository extends JpaRepository<EventRequest, Long>, EventRequestRepositoryCustom {
+public interface EventRequestRepository extends JpaRepository<EventRequest, Integer>, EventRequestRepositoryCustom {
 }
