+++ src/main/java/compedia/vn/tickmi/mail/repository/TicketRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket, Long>, TicketRepositoryCustom {
+public interface TicketRepository extends JpaRepository<Ticket, Integer>, TicketRepositoryCustom {
 }
